\name{dreport-package}
\alias{dreport-package}
\alias{dreport}
\docType{package}
\title{
\packageTitle{dreport}
}
\description{
\packageDescription{dreport}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dreport}
\packageIndices{dreport}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{dreport}

Maintainer: \packageMaintainer{dreport}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
